fift: Do not push bool on stack in *@ without ?
Functions like `i@` (without `?`) push `false` on stack in case of
error. I am not sure if it matters at all as, from my experience, in
case of an exception the stack gets cleared, however, I imagine, it
might become important with custom exception handlers.

Make sure the error code is only ever pushed if the function name had `?`
in it.