Introduce vm::CellStorageStatComputer for calculating CSS
The commit marks CSSComputer as non-copyable and non-movable and makes
CSS a simple struct consisting of 3 integer fields, eliminating the need
for awkward CSS::clear{,_seen,_limit} calls and ensuring that implicit
copy constructors for CSS are always cheap.

By doing so, the patch improves performance on the verification
benchmark by about 1% by removing an accidental (and implicit!) copy of
a large hash map.