Avoid using vm::load_cell_slice_special in AccountStorageStat
Block verification tests we have currently omit cells that were not
visited by the old cell storage fee computer due to exceeding
predetermined limits. If run on such subtrees before this commit, new
implementation threw an exception that was later reinterpreted as block
failing verification.

The patch corrects this by not throwing an exception in the first place.