Teach full node to response with split persistent states
We introduce two additional ADNL requests that can be made between
nodes: downloadPersistentStateSliceV2 and getPersistentStateSizeV2. They
differ from "V1" versions by inclusion of "effective shard" parameter.
The parameter denotes a split account state if block.shard is a proper
ancestor of effective_shard. If block.shard == effective_shard, we treat
this as a request for split state header, and if it is 0 (or invalid),
we serve an unsplit state.

Old requests are wired to call V2 with effective shard set to 0.