Remove concept of virtualization level (#1648)
* Mostly do not reference virtualization level outside of cells

This is done by adding `Cell::is_virtualized` method that checks if
`virtualization != 0` and providing a default value for
vm::MerkleProof::virtualize virtualization parameter. Everywhere except
for a single place in tests (where it doesn't matter), the parameter was
set to 1.

* Remove concept of virtualization level

We use virtualization level 0 for non-virtualized cells and level 1 for
virtualized ones. With this patch, we do not store this level anymore
and instead infer it based on "effective" level (previously known as
VirtualizationParameters::level) and the actual level of the cell.

This is mainly done to reduce cognitive complexity of the code. However,
as a nice side-effect, changes in VirtualCell give a negligible
performance improvement on the verification benchmark.